ansible/requirements.txt
ansible >2.9.0,<2.10
openstacksdk
pyyaml
